+++ index.d.ts
@@ -7,6 +7,7 @@ export {
   REPLACE,
   THROW,
   default as insert,
+  insertByKey,
   insertIfNotExists,
   insertOrReplaceIfExists,
   insertOrThrowIfExists,